Prevent "make dist" from running autoreconf
We include the auto-generated configure and config.h.in files in the
repository in order to avoid infection of GNU GPL due to inclusion of
config.guess and config.sub, which are distributed under GPL with a
special exception.  This lets us maintain configure and config.h.in in
the repository as if they are generated from the latest configure.ac.
Concequently, we do not need to regenerate them at the timing of
packaging as their latest version is in the repository.